fix(wpp-open-agent-updater): Address bugs found during 5-agent code review
- Fix @CurrentUser() returning full object instead of userId string
- Prevent duplicate runs by checking both PENDING and PROCESSING status
- Change fileSize column from bigint to int (pg returns bigint as string)
- Add @IsOptional() to wppOpenToken in CreateTaskDto
- Add network error handling and null guards in WPP Open service
- Only set Content-Type header when request body is present
- Fix data inconsistency when knowledge upsert fails (bulk update files)
- Add nested try/catch in worker error handler to prevent crashes
- Log rejected subfolder scans in Box recursive traversal
- Fix oversized file status/counter mismatch (failed not skipped)
- Add missing error handlers in 3 frontend components
- Remove unnecessary `as any` casts and unused TableModule import
- Add missing barrel exports for AGENT_UPDATER_QUEUE and job data type
- Add query parameter validation on listAgents endpoint

+++ apps/api/migrations/1709836900000-CreateWppOpenAgentUpdaterSchema.ts
@@ -70,7 +70,7 @@ export class CreateWppOpenAgentUpdaterSchema1709836900000
 				"taskRunId" uuid NOT NULL,
 				"boxFileId" character varying(255) NOT NULL,
 				"fileName" character varying(500) NOT NULL,
-				"fileSize" bigint NOT NULL DEFAULT 0,
+				"fileSize" integer NOT NULL DEFAULT 0,
 				"status" "wpp_open_agent_updater"."task_run_files_status_enum" NOT NULL DEFAULT 'pending',
 				"errorMessage" text,
 				"processedAt" TIMESTAMP WITH TIME ZONE,
